refactors code
