+++ src/components/App/App.js
@@ -1,21 +1,21 @@
 import React, { useEffect } from 'react';
 import jService from '../../api/jService';
 import filterCluesWithValues from '../../logic/getCategoryClues';
-
-import generateCategoryIdSet from '../../logic/logic';
+import generateCategoryIdSet from '../../logic/generateCategoryIdSet';
 
 function App() {
   useEffect(() => {
     getClues();
   }, []);
 
-  const getClues = async () => {
+  const getClues = async (categoryID) => {
     const { data } = await jService.get('/', {
       params: {
         id: ids[0],
       },
     });
-
+    // getClues will need to be called for each category column.
+    // It will need a categoryID which is generated by generateCategoryIdSet().
     return console.log(filterCluesWithValues(data.clues, data.title));
   };
