+++ src/logic/generateCategoryIdSet.js
